original energy-optimizer/requirements.txt
# Antigravity Energy Optimizer - Dependencies
# Python 3.10+

# =============================================================================
# Core Framework
# =============================================================================
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# =============================================================================
# Data Processing
# =============================================================================
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# =============================================================================
# Visualization
# =============================================================================
plotly>=5.18.0
altair>=5.0.0
folium>=0.14.0
streamlit-folium>=0.15.0

# =============================================================================
# Optimization (Pyomo + Solvers)
# =============================================================================
pyomo>=6.6.0
# Note: Install solver separately:
# - Gurobi: Commercial, best performance (pip install gurobipy, needs license)
# - CBC: Free, good for development (conda install -c conda-forge coincbc)
# - GLPK: Free, basic (conda install -c conda-forge glpk)

# =============================================================================
# Backend / Data Storage
# =============================================================================
# Google Sheets
gspread>=5.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# SharePoint (future)
# msal>=1.24.0
# Office365-REST-Python-Client>=2.5.0

# =============================================================================
# Export / Reporting
# =============================================================================
python-pptx>=0.6.21
openpyxl>=3.1.0
reportlab>=4.0.0
jinja2>=3.1.0

# =============================================================================
# AI Integration (Optional)
# =============================================================================
# google-generativeai>=0.3.0  # For Gemini
# anthropic>=0.7.0            # For Claude

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2023.3
